try to better handle the "health would like to send you notifications" alert
