+++ Sources/XCTHealthKit/XCTest+AddSamples.swift
@@ -53,74 +53,4 @@ extension XCTestCase {
             }
         }
     }
-    
-    
-    /// Walks through the Health App onboarding flow, if necessary.
-    @MainActor
-    public func handleHealthAppOnboardingIfNecessary(_ healthApp: XCUIApplication) {
-        if healthApp.staticTexts["Welcome to Health"].waitForExistence(timeout: 3) {
-            handleOnboarding(healthApp)
-        }
-    }
-    
-    
-    @MainActor
-    func handleOnboarding(_ healthApp: XCUIApplication, alreadyRecursive: Bool = false) {
-        self.addUIInterruptionMonitor(withDescription: "System Dialog") { alert in
-            guard alert.buttons["Allow"].exists else {
-                XCTFail("Failed not dismiss alert: \(alert.staticTexts.allElementsBoundByIndex)")
-                return false
-            }
-            
-            alert.buttons["Allow"].tap()
-            return true
-        }
-        
-        if healthApp.staticTexts["Welcome to Health"].waitForExistence(timeout: 5) {
-            XCTAssertTrue(healthApp.staticTexts["Continue"].waitForExistence(timeout: 5))
-            healthApp.staticTexts["Continue"].tap()
-            
-            XCTAssertTrue(healthApp.staticTexts["Continue"].waitForExistence(timeout: 5))
-            healthApp.staticTexts["Continue"].tap()
-            
-            XCTAssertTrue(healthApp.tables.buttons["Next"].waitForExistence(timeout: 5))
-            healthApp.tables.buttons["Next"].tap()
-            
-            // Sometimes the HealthApp fails to advance to the next step here.
-            // Go back and try again.
-            if !healthApp.staticTexts["Continue"].waitForExistence(timeout: 60) {
-                // Go one step back.
-                healthApp.navigationBars["WDBuddyFlowUserInfoView"].buttons["Back"].tap()
-                
-                XCTAssertTrue(healthApp.staticTexts["Continue"].waitForExistence(timeout: 5))
-                healthApp.staticTexts["Continue"].tap()
-                
-                // Check if the Next button exists or of the view is still in a loading process.
-                if healthApp.tables.buttons["Next"].waitForExistence(timeout: 5) {
-                    healthApp.tables.buttons["Next"].tap()
-                }
-                
-                // Continue button still doesn't exist, go for terminating the app.
-                if !healthApp.staticTexts["Continue"].waitForExistence(timeout: 60) {
-                    if alreadyRecursive {
-                        logger.notice("Even the recursive process did fail. Terminate the process.")
-                    }
-                    
-                    healthApp.terminate()
-                    healthApp.activate()
-                    handleOnboarding(healthApp, alreadyRecursive: true)
-                    return
-                }
-            }
-            
-            // Try to turn off the Health Notifications Trends Switch:
-            let trendsSwitch = healthApp.switches.firstMatch
-            if trendsSwitch.waitForExistence(timeout: 5) && trendsSwitch.isHittable {
-                trendsSwitch.tap()
-            }
-            
-            XCTAssertTrue(healthApp.staticTexts["Continue"].waitForExistence(timeout: 5))
-            healthApp.staticTexts["Continue"].tap()
-        }
-    }
 }
